Moved create to users controller
Refactoring, moving everything to the userscontroller
