+++ app/Http/Controllers/Admin/TrainerController.php
@@ -16,89 +16,7 @@ class TrainerController extends Controller
 {
 
 
-    function generateRandomString($length = 10) {
-        $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
-        $charactersLength = strlen($characters);
-        $randomString = '';
-        for ($i = 0; $i < $length; $i++) {
-            $randomString .= $characters[rand(0, $charactersLength - 1)];
-        }
-        return $randomString;
-    }
 
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        $request->validate([
-
-            'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
-            'lastname' => ['required'],
-            'birthdate' => ['required'],
-            'code' => ['required'],
-            'phonenumber' => ['required'],
-            'city' => ['required'],
-            'street' => ['required'],
-            'number' => ['required'],
-            'post_code' => ['required'],
-            'work_start' => ['required'],
-            'end_work' => ['required'],
-            'state' => ['required'],
-            'experience' => ['required'],
-
-        ]);
-        $pass = $this->generateRandomString();
-        $a = Address::create([
-            'street' => $request->street,
-            'city' => $request->city,
-            'number' => $request->number,
-            'post_code' => $request->post_code,
-
-        ]);
-        $trainerRole = Role::where('name', 'trainer')->first()->id;
-        $user = User::create([
-            'role_id' => $trainerRole,
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($pass),
-            'lastname' => $request->lastname,
-            'phonenumber' => $request->phonenumber,
-            'code' => $request->code,
-            'birthdate' => $request->birthdate,
-            'address_id' => $a->id
-        ]);
-        Trainer::create([
-            'user_id' => $user->id,
-            'work_start' =>$request->work_start,
-            'end_work' => $request->end_work,
-            'state_id' => $request->state,
-            'experience' => $request->experience
-        ]);
-        return redirect()->route('admin.users.index');
-    }
-
-
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Trainer  $trainer
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Trainer $trainer)
-    {
-        $states = State::all();
-        return view('admin.trainers.edit')->with([
-            'trainer' => $trainer,
-            'states' => $states
-        ]);
-    }
 
     /**
      * Update the specified resource in storage.
