+++ app/Http/Controllers/Admin/UserController.php
@@ -3,15 +3,28 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Address;
+use App\Models\Admin;
 use App\Models\State;
+use App\Models\Trainer;
 use App\Models\User;
 use App\Models\Role;
 use Illuminate\Http\Request;
 use Gate;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+    function generateRandomString($length = 10) {
+        $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+        $charactersLength = strlen($characters);
+        $randomString = '';
+        for ($i = 0; $i < $length; $i++) {
+            $randomString .= $characters[rand(0, $charactersLength - 1)];
+        }
+        return $randomString;
+    }
     public function search(Request $request){
 
         $uzklausa = '%'.$request->ieskoti.'%';
@@ -28,6 +41,67 @@ class UserController extends Controller
         $states = State::all();
         return view('admin.users.create')->with('states', $states);
     }
+    public function store(Request $request)
+    {
+        $request->validate([
+
+            'name' => ['required', 'string', 'max:255'],
+            'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
+            'lastname' => ['required'],
+            'birthdate' => ['required'],
+            'code' => ['required'],
+            'phonenumber' => ['required'],
+            'city' => ['required'],
+            'street' => ['required'],
+            'number' => ['required'],
+            'post_code' => ['required'],
+            'state' => ['required'],
+        ]);
+        $pass = $this->generateRandomString();
+        $trainerRole = Role::where('name', 'trainer')->first()->id;
+        $adminRole = Role::where('name', 'admin')->first()->id;
+        $a = Address::create([
+            'street' => $request->street,
+            'city' => $request->city,
+            'number' => $request->number,
+            'post_code' => $request->post_code,
+
+        ]);
+        $user = User::create([
+            'role_id' => $trainerRole,
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => Hash::make($pass),
+            'lastname' => $request->lastname,
+            'phonenumber' => $request->phonenumber,
+            'code' => $request->code,
+            'birthdate' => $request->birthdate,
+            'address_id' => $a->id
+        ]);
+        if($request->role==='trainer')
+        {
+            Trainer::create([
+                'user_id' => $user->id,
+                'work_start' =>$request->work_start,
+                'end_work' => $request->end_work,
+                'state_id' => $request->state,
+                'experience' => $request->experience
+            ]);
+        }
+        else{
+            $user->role_id=$adminRole;
+            $user->save();
+            Admin::create([
+                'user_id' => $user->id,
+                'work_start' =>$request->work_start,
+                'end_work' => $request->end_work,
+                'state_id' => $request->state
+            ]);
+        }
+
+        return redirect()->route('admin.users.index');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -37,6 +111,30 @@ class UserController extends Controller
     {
         $users = User::all();
         return view('admin.users.index')->with('users', $users);
+    }
+    public function edit(User $user)
+    {
+        $states = State::all();
+        if($user->hasRole('admin'))
+        {
+            return view('admin.admins.edit')->with([
+                'user' => $user,
+                'states' => $states
+            ]);
+        }
+        else if ($user->hasRole('trainer'))
+        {
+            $states = State::all();
+            return view('admin.trainers.edit')->with([
+                'user' => $user,
+                'states' => $states
+            ]);
+        }
+        else if ($user->hasRole('user'))
+        {
+            return view('admin.clients.edit')->with('user',$user);
+        }
+
     }
